+++ src/Sulu/Bundle/LocationBundle/Tests/app/AppKernel.php
@@ -6,19 +6,4 @@ use Sulu\Bundle\TestBundle\SuluTestBundle;
 
 class AppKernel extends SuluTestKernel
 {
-    public function registerBundles()
-    {
-        return array_merge(
-            parent::registerBundles(),
-            array(
-                new \Sulu\Bundle\LocationBundle\SuluLocationBundle(),
-            )
-        );
-    }
-
-    public function registerContainerConfiguration(LoaderInterface $loader)
-    {
-        $loader->load(SuluTestBundle::getConfigDir() . '/config.php');
-        $loader->load(__DIR__ . '/config/config.yml');
-    }
 }