+++ src/Sulu/Bundle/MediaBundle/Content/MediaSelectionContainer.php
@@ -17,7 +17,6 @@ use JMS\Serializer\Annotation\Exclude;
 
 /**
  * Container for Image selection, holds config for image selection and lazy loads images matches the ids
- * @package Sulu\Bundle\MediaBundle\Content
  */
 class MediaSelectionContainer implements ArrayableInterface
 {