+++ CHANGELOG.md
@@ -6,13 +6,19 @@ CHANGELOG for Sulu
     * BUGFIX      #846 [MediaBundle]     Added missing dot to create event name method (\cc Daniel)
     * ENHANCEMENT #841 [SecurityBundle]  Unique email per user
     * BUGFIX      #837 [AdminBundle]     Javascript function for croping labels with a certain tag this.sandbox.sulu.cropAllLabels(className)
+    * ENHANCEMENT #818 [ContentBundle]   Enhanced column-navigation ordering ui
+    * BUGFIX      #857 [ContentBundle]   Added links without save could not be removed
+    * FEATURE     #789 [ContentBundle]   Added present as to smart content config (see [here ...](https://github.com/sulu-cmf/docs/blob/master/developer-documentation/300-webspaces/smart-content.md))
+    * BUGFIX      #856 [ContentBundle]   Added default values for smart content view vars
+
+* 0.15.2 (2015-02-19)
+    * HOTFIX      #850 [MediaBundle]     Fixed bug with deleted media in media selection
 
 * 0.15.1 (2015-02-17)
     * HOTFIX      #842 [CoreBundle]      Fixed upgrade internal links command for installations without snippets
 
 * 0.15.0 (2015-02-13)
     * BUGFIX      #833 [AdminBundle]     Added new husky version
-    * ENHANCEMENT #818 [ContentBundle]   Enhanced column-navigation ordering ui
     * BUGFIX      #829 [ContactBundle]   Account-Contacts: show full-name of contact
     * ENHANCEMENT #828 [ContactBundle]   Changed columns for contact list and made concatenated columns not sortable
     * BUGFIX      #825 [WebsiteBundle]   Fixed syntax error in ExceptionController
@@ -39,6 +45,7 @@ CHANGELOG for Sulu
     * BUGFIX      #778 [ContentBundle]   Fixed shadow page with internal link and smart-content
     * BUGFIX      #790 [WebsiteBundle]   Fixed twig variables for 404 page
     * FEATURE     #791 [ContentBundle]   Added Changelog to settings tab
+    * FEATURE     #789 [ContentBundle]   Enabled property parameters to have metadata for localization
     * FEATURE     #684 [ContentBundle]   Refactored preview to use new websocket component and only one socket for form
                                          and preview
     * FEATURE     #684 [WebsocketBundle] Implemented Websocket Component to standardize Websocket implementations